For large downloads, direct to disk, the length field has to be more than 4 bytes. This is the length field in struct i_get. This is systemic, and has ramifications. Example I had to make 8 byte wrappers for all the string functions, as they often operate on g.buffer and g.length. Eventually a lot more variables have to become 8 bytes, and maybe these new string functions will become the standard some day, but I couldn't change everything all at once. We have to modify the ship while it is still sailing.

+++ src/eb.h
@@ -174,7 +174,7 @@ extern const char *ebrc_string; /* default ebrc file */
 struct i_get {
 // the data returned from the internet fetch
 	char *buffer;
-	int length;
+	long long length;
 // in case you want the headers
 	char **headers_p;
 	const char *url;
@@ -358,7 +358,7 @@ extern char *emojiFile;	// your library of emojis
 extern char *irclog; // log all your irc sessions
 extern int redirect_count;
 extern char *serverData;
-extern int serverDataLen;
+extern long long serverDataLen;
 extern char *breakLineResult;
 extern char *home;		/* home directory */
 extern char *recycleBin;	/* holds deleted files */
